Updated dependencies, removed Instagram, use Google API wrapper
Several dependencies we were using either no longer exist or have new names.
Instagram no longer supports an API that will let us fetch photos in the way we
were doing, so we can't support that anymore. The Google Drive gem we were
using no longer exists so I tried to do what we were doing with a lower-level
Google API wrapper.

This works... sort of. For even minimal use, it will quickly exceed Google's API
rate limits, so it needs to be refactored to fetch all the data from the sheet,
make modifications locally, then commit them to the sheet in a single request.